Added TimeCountDown and motivation+inspiration feature.

+++ CarpeDiem.Core/TimeManagement.vb
@@ -1,15 +1,15 @@
 ﻿Public Class TimeManagement
 
 
-    Public Shared Function TimeAfterHours(hours As Integer) As Date
+    Public Shared Function TimeAfterHours(hours As Double) As Date
         Return Date.Now.AddHours(hours)
     End Function
 
-    Public Shared Function TotalMinutesAfter(hours As Integer) As Integer
+    Public Shared Function TotalMinutesAfter(hours As Double) As Integer
         Return TimeAfterHours(hours).Subtract(Date.Now).TotalMinutes
     End Function
 
-    Public Shared Function GetTimeAfterMinutes(minutes As Integer) As Date
+    Public Shared Function GetTimeAfterMinutes(minutes As Double) As Date
         Return now.AddMinutes(minutes)
     End Function
